Moves more general function with other general functions
